# [Performance] Improve Character Select DB Performance