# Description

This is a performance adjustment for **World** where on THJ at 4,000 players and 2,000 zoneservers we observe more of the stress points of world. This cuts down on query volume dramatically.

When you have 12 characters, you can produce well over 100 queries just by hitting character select. The same amount of queries get produced when you delete or create a character while in character select. Most of these queries are from queries inside of a loop. Instead, we bulk load as much as we can at once and only run about 5 queries.

[Example](https://gist.github.com/Akkadius/34e0639c5f27871cee16b0355874d89d) of World queries before when someone would hit character select (Previously)

**Bulk Loading**

* We fetch all character rows at once
* We fetch all databuckets at once
* We fetch all inventory rows for all characters at once
* We fetch all binds for all characters at once
* We fetch all inventory material colors at once
* We no longer do an extra query to get character id by name since we already have it

## Type of change

- [x] Optimization

# Testing

* Character creation, character goes to the expected bind
* Character deletion, character gets soft deleted and shown as expected
* Character inventory shows as expected
* Character inventory material colors shows as expected

Only 5 queries are made during character select, no matter how many characters I have

```
 World |   Query    | QueryDatabase SELECT id, account_id, name, last_name, title, suffix, zone_id, zone_instance, y, x, z, heading, gender, race, `class`, level, deity, birthday, last_login, time_played, level2, anon, gm, face, hair_color, hair_style, beard, beard_color, eye_color_1, eye_color_2, drakkin_heritage, drakkin_tattoo, drakkin_details, ability_time_seconds, ability_number, ability_time_minutes, ability_time_hours, exp, exp_enabled, aa_points_spent, aa_exp, aa_points, group_leadership_exp, raid_leadership_exp, group_leadership_points, raid_leadership_points, points, cur_hp, mana, endurance, intoxication, str, sta, cha, dex, `int`, agi, wis, extra_haste, zone_change_count, toxicity, hunger_level, thirst_level, ability_up, ldon_points_guk, ldon_points_mir, ldon_points_mmc, ldon_points_ruj, ldon_points_tak, ldon_points_available, tribute_time_remaining, career_tribute_points, tribute_points, tribute_active, pvp_status, pvp_kills, pvp_deaths, pvp_current_points, pvp_career_points, pvp_best_kill_streak, pvp_worst_death_streak, pvp_current_kill_streak, pvp2, pvp_type, show_helm, group_auto_consent, raid_auto_consent, guild_auto_consent, leadership_exp_on, RestTimer, air_remaining, autosplit_enabled, lfp, lfg, mailkey, xtargets, firstlogon, e_aa_effects, e_percent_to_aa, e_expended_aa_spent, aa_points_spent_old, aa_points_old, e_last_invsnapshot, UNIX_TIMESTAMP(deleted_at), illusion_block FROM character_data WHERE `account_id` = 2454 AND `deleted_at` IS NULL ORDER BY `name` LIMIT 12 -- (10 rows returned) (0.000868s)
 World |   Query    | QueryDatabase SELECT character_id, slot_id, item_id, charges, color, augment_one, augment_two, augment_three, augment_four, augment_five, augment_six, instnodrop, custom_data, ornament_icon, ornament_idfile, ornament_hero_model, guid FROM inventory WHERE `character_id` IN (7129,43465,43462,43459,43458,43463,43460,43461,43464,43467) AND `slot_id` BETWEEN 2 AND 19 -- (163 rows returned) (0.000687s)
 World |   Query    | QueryDatabase SELECT id, slot, zone_id, instance_id, x, y, z, heading FROM character_bind WHERE `id` IN (7129,43465,43462,43459,43458,43463,43460,43461,43464,43467) ORDER BY id, slot -- (50 rows returned) (0.000461s)
 World |   Query    | QueryDatabase SELECT id, slot, blue, green, red, use_tint, color FROM character_material WHERE `id` IN (7129,43465,43462,43459,43458,43463,43460,43461,43464,43467) ORDER BY id, slot -- (14 rows returned) (0.000140s)
 World |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE `key` = 'GestaltClasses' AND character_id IN (7129,43465,43462,43459,43458,43463,43460,43461,43464,43467) -- (10 rows returned) (0.000502s)
```

![image](https://github.com/user-attachments/assets/f5710942-6c67-4378-94b2-6b27784bb08e)

![image](https://github.com/user-attachments/assets/1a483f79-5d5b-4796-b341-50055e03ca44)

Grabbed Wizard in game, reflected properly in character select

![image](https://github.com/user-attachments/assets/4f449f1f-69d9-42be-b01f-a49377443240)

# Checklist

- [x] I have tested my changes
- [x] I have performed a self-review of my code. Ensuring variables, functions and methods are named in a human-readable way, comments are added only where naming of variables, functions and methods can't give enough context.
- [x] I own the changes of my code and take responsibility for the potential issues that occur